+++ src/main/java/edu/ie3/datamodel/io/sink/CsvFileSink.java
@@ -334,8 +334,8 @@ public class CsvFileSink implements InputDataSink, OutputDataSink {
   }
 
   /**
-   * Transforms a provided map of string to string to valid csv formatted strings (according to csv
-   * specification RFC 4180)
+   * Transforms a provided map of string to string into valid csv formatted strings (according to
+   * csv specification RFC 4180)
    *
    * @param entityFieldData a string to string map that should be processed
    * @return a new map with valid csv formatted keys and values strings