+++ src/test/resources/edu/ie3/datamodel/io/sink/_sql/input_entities.sql
@@ -41,7 +41,7 @@ CREATE TABLE public.evcs_input
     controlling_em UUID,
     cos_phi_rated TEXT NOT NULL,
     id TEXT NOT NULL,
-    location_type TEXT NOT NULL,
+    location_types TEXT NOT NULL,
     node UUID NOT NULL,
     operates_from TIMESTAMP WITH TIME ZONE,
     operates_until TIMESTAMP WITH TIME ZONE,